# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vacancies API",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/api": {},
        "/api/vacancies": {
            "get": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Fetch vacancies",
                "operationId": "7ec6e87c684b8facfe9b93a3ad9e40ef",
                "parameters": [
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Dynamic fields. Can be string(fields=a, b, c) or array(fields[]=a, b, c)",
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "null": {
                                "summary": "",
                                "value": ""
                            },
                            "created_at": {
                                "summary": "With column created_at.",
                                "value": "created_at"
                            },
                            "updated_at": {
                                "summary": "With column updated_at.",
                                "value": "updated_at"
                            },
                            "requirements": {
                                "summary": "With column requirements.",
                                "value": "requirements"
                            },
                            "with all": {
                                "summary": "With all columns.",
                                "value": "created_at, updated_at, requirements"
                            }
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column",
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "null": {
                                "summary": "Not provided",
                                "value": ""
                            },
                            "created_at": {
                                "summary": "Sort by created_at column.",
                                "value": "created_at"
                            },
                            "salary": {
                                "summary": "Sort by salary column.",
                                "value": "salary"
                            }
                        }
                    },
                    {
                        "name": "sort_type",
                        "in": "query",
                        "description": "Sort by column",
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "null": {
                                "summary": "If not provided use default ASC.",
                                "value": ""
                            },
                            "asc": {
                                "summary": "Sort ASC.",
                                "value": "asc"
                            },
                            "desc": {
                                "summary": "Sort DESC.",
                                "value": "desc"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": [
                                                {
                                                    "id": 1,
                                                    "name": "Some name",
                                                    "description": "Some description.",
                                                    "salary": 120000
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "Some name 2",
                                                    "description": "Some description 2.",
                                                    "salary": 10000
                                                }
                                            ],
                                            "links": {
                                                "first": "http://localhost:8000/api/vacancies?page=1",
                                                "last": "http://localhost:8000/api/vacancies?page=4",
                                                "prev": null,
                                                "next": "http://localhost:8000/api/vacancies?page=2"
                                            },
                                            "meta": {
                                                "current_page": 1,
                                                "from": 1,
                                                "last_page": 4,
                                                "links": [
                                                    {
                                                        "url": null,
                                                        "label": "&laquo; Previous",
                                                        "active": false
                                                    },
                                                    {
                                                        "url": "http://localhost:8000/api/vacancies?page=1",
                                                        "label": "1",
                                                        "active": true
                                                    },
                                                    {
                                                        "url": "http://localhost:8000/api/vacancies?page=2",
                                                        "label": "2",
                                                        "active": false
                                                    },
                                                    {
                                                        "url": "http://localhost:8000/api/vacancies?page=3",
                                                        "label": "3",
                                                        "active": false
                                                    },
                                                    {
                                                        "url": "http://localhost:8000/api/vacancies?page=4",
                                                        "label": "4",
                                                        "active": false
                                                    },
                                                    {
                                                        "url": "http://localhost:8000/api/vacancies?page=2",
                                                        "label": "Next &raquo;",
                                                        "active": false
                                                    }
                                                ],
                                                "path": "http://localhost:8000/api/vacancies",
                                                "per_page": 10,
                                                "to": 10,
                                                "total": 32
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "params with wrong type|value.",
                                        "value": {
                                            "message": "The selected sort by is invalid.",
                                            "errors": {
                                                "sort_by": [
                                                    "The selected sort by is invalid."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Add new vacancy",
                "operationId": "342e75f8702d3c5141715fa10e7f90d7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Test name"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Test description"
                                            },
                                            "requirements": {
                                                "type": "string",
                                                "example": "Test requirements"
                                            },
                                            "salary": {
                                                "type": "integer",
                                                "example": "100000"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "description": "ID of the sotred row",
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "no required parameters.",
                                        "value": {
                                            "message": "The description field is required. (and 1 more error).",
                                            "errors": {
                                                "description": [
                                                    "The description field is required."
                                                ],
                                                "requirements": [
                                                    "The requirements field is required."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vacancies/{id}": {
            "put": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "update vacancy",
                "operationId": "5c50fabb40f9989f935d6be293de96d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "New name",
                                "value": "new vacancy name"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "ok": {
                                        "summary": "Example.",
                                        "value": {
                                            "status": "ok",
                                            "data": 1
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Example.",
                                        "value": {
                                            "status": "error",
                                            "data": "no query results with id = 2"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Example.",
                                        "value": {
                                            "message": "The salary field must be an integer.",
                                            "errors": {
                                                "salary": [
                                                    "The salary field must be an integer."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Delete vacancy",
                "operationId": "e867e296e4b4e17d93668644cf5bc577",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "ok": {
                                        "summary": "Example.",
                                        "value": {
                                            "status": "ok",
                                            "data": 1
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Example.",
                                        "value": {
                                            "status": "error",
                                            "data": "no query results with id = 2"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Vacancies",
            "description": "Vacancies"
        }
    ]
}